requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-fs-trumbowyg==0.1.4
django-js-asset==1.2.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
idna==2.10
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
gunicorn
dj_database_url
